Add model_regression workflow
